KGA: A Hero's Beginning
So this is my first game I made in Game Maker Studio 2. I only have the trial version so I can't export an executable version of the game. But there's another way: If u have GM Studio 2 u can open the project file and then start the game via GM Studio 2. Sorry that it's so complicated but I don't have the money for Game Maker Studio 2. 

As probably saw in the title, the game is only in Ver 0.1.  The game isn't finished at all but I still wanted to upload it.

-there is only one room
-u can jump
-u can do a boost jump
-sprint
-there are walk animations
-i added enemys 
-i added a gun
  -u can shoot the enemys
    -after 4 hits they're dead

So I will continue to work on this game ;)